feat: eliminar SelectInvoice y crear StepInvoice para gestionar la selección de facturas
